// src/app/features/casino/models/roulette2-data.model.ts
export interface Roulette2Data {
  straightUpMultiplicator: number;
  splitMultiplicator: number;
  streetMultiplicator: number;
  cornerMultiplicator: number;
  sixLineMultiplicator: number;
  columnMultiplicator: number;
  dozenMultiplicator: number;
  oddEvenMultiplicator: number;
  redBlackMultiplicator: number;
  middleMultiplicator: number;
  maxBet: number;
  vipStraightUpMultiplicator: number;
  vipSplitMultiplicator: number;
  vipStreetMultiplicator: number;
  vipCornerMultiplicator: number;
  vipSixLineMultiplicator: number;
  vipColumnMultiplicator: number;
  vipDozenMultiplicator: number;
  vipOddEvenMultiplicator: number;
  vipRedBlackMultiplicator: number;
  vipMiddleMultiplicator: number;
  vipMaxBet: number;
}